Implement ARPEGGIO_FAST plus fixes
Fixed start/end of arpeggio. The initial logic was wrong and the Arpeggio notes did not always consist of four notes as a result.

Also removed audible glitch when transitioning out of Arpeggio by letting peekNextNote return the correct next note also in this case.

Also halved number of Arpeggio notes when tune-speed is very fast (to match PICO-8 behaviour).