Changed NOISE2 implementation
It is now based on a SQUARE wave instead of a TRIANGLE wave.
Four times each wave-cycle, the index delta is randomly changed.
This way, the relative wave-form distortion is still the same, irrespective of the note frequency.
However, the noise sounds far less harsh then the original square noise.