Improve blending
Incompatible notes now blend via LERP-ing.
This way, the wave forms from both notes contribute equally to the transition.
This fixes the glitches that were present in the "The Lair" song.

The drawback of the new algorithm is that it requires extra CPU, as during the transition the samples for both notes are calculated.
Furthermore, the blending uses buffers, so the memory footprint has also gone up slightly.